[UPDATE] Update snake movement
I update the snake movement to add the body and upgrade the body when an apple is consume.
The body upgrade from the tail progressively.

The snake sprite can be modified. It just need to correspond to a 32x32 files.

I also divided the code in different file as recommended.